Controller cannot access Viewports (main.js does)
