+++ index.html
@@ -58,6 +58,6 @@
 	<script src='Scripts/webgl-obj-loader.js'></script>
 	<script src="http://d3js.org/d3.v3.min.js" charset="utf-8"></script>
 	<script type="text/javascript" src="Scripts/jquery-3.3.1.js"></script>
-  <script type="module" src="Scripts/controller.js"></script>
+  <script type="module" src="Scripts/main.js"></script>
 
 </html>
